requirements.txt
semantic-kernel
openapi-schema-pydantic
rich
python-dotenv